WIP: Experimenting with ImageJ1-based segmentation and measurement macros headless
It's tough because ImageJ1 doesn't work headless, and in particular its
WindowManager tracks ImageWindows, which don't exist headless, so you
need to bend over backwards to register images with ImageJ1.
Hopefully this notebook helps to illustrate what works and what doesn't.

BUT THIS WIP DOES NOT YET CLEARLY ILLUSTRATE THE CHALLENGES. STAY TUNED.